+++ src/main/java/my_social_media_project_backend/demo/custom/CustomUserDetails.java
@@ -6,14 +6,14 @@ import org.springframework.security.core.userdetails.User;
 import java.util.Collection;
 
 public class CustomUserDetails extends User {
-    private final Integer userId;
+    private final Long userId;
 
-    public CustomUserDetails(Integer userId, String accountName, String password, Collection<? extends GrantedAuthority> authorities) {
+    public CustomUserDetails(Long userId, String accountName, String password, Collection<? extends GrantedAuthority> authorities) {
         super(accountName, password, authorities);
         this.userId = userId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 }
