+++ src/main/java/my_social_media_project_backend/demo/dto/FriendshipDTO.java
@@ -5,23 +5,23 @@ import my_social_media_project_backend.demo.entity.Friendship;
 import java.time.LocalDateTime;
 
 public class FriendshipDTO {
-    private final Integer userId;
-    private final Integer friendId;
+    private final Long userId;
+    private final Long friendId;
     private final Friendship.Status status;
     private final LocalDateTime createAt;
 
-    public FriendshipDTO(Integer userId, Integer friendId, Friendship.Status status, LocalDateTime createAt) {
+    public FriendshipDTO(Long userId, Long friendId, Friendship.Status status, LocalDateTime createAt) {
         this.userId = userId;
         this.friendId = friendId;
         this.status = status;
         this.createAt = createAt;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public Integer getFriendId() {
+    public Long getFriendId() {
         return friendId;
     }
