+++ src/main/java/my_social_media_project_backend/demo/dto/PostLikeDTO.java
@@ -1,13 +1,13 @@
 package my_social_media_project_backend.demo.dto;
 
 public class PostLikeDTO {
-    private final Integer user_id;
+    private final Long user_id;
 
-    public PostLikeDTO(Integer user_id) {
+    public PostLikeDTO(Long user_id) {
         this.user_id = user_id;
     }
 
-    public Integer getUser_id() {
+    public Long getUser_id() {
         return user_id;
     }
 }
