+++ src/main/java/my_social_media_project_backend/demo/dto/PostWithUserDTO.java
@@ -8,7 +8,7 @@ public class PostWithUserDTO extends PostDTO{
 
     private final UserDTO user;
 
-    public PostWithUserDTO(Integer id, String title, String content, long likeCount, long commentCount, boolean isLiked, LocalDateTime create_at, UserDTO user) {
+    public PostWithUserDTO(Long id, String title, String content, long likeCount, long commentCount, boolean isLiked, LocalDateTime create_at, UserDTO user) {
         super(id, title, content, likeCount, commentCount, isLiked, create_at);
         this.user = user;
     }
