+++ src/main/java/my_social_media_project_backend/demo/entity/Friendship.java
@@ -14,7 +14,7 @@ public class Friendship {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "status", nullable = false)
@@ -32,10 +32,10 @@ public class Friendship {
     private User friend;
 
     @Column(name = "user_id", nullable = false)
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "friend_id", nullable = false)
-    private Integer friendId;
+    private Long friendId;
 
     public long getId() {
         return id;
@@ -77,19 +77,19 @@ public class Friendship {
         this.friend = friend;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getFriendId() {
+    public Long getFriendId() {
         return friendId;
     }
 
-    public void setFriendId(Integer friendId) {
+    public void setFriendId(Long friendId) {
         this.friendId = friendId;
     }
 }
